+++ src/interfaces/TemplateConfig.ts
@@ -1,0 +1,8 @@
+export interface TemplateConfig {
+  button: SceneControlTool;
+  data: MeasuredTemplateConstructorDataData;
+}
+
+export type MeasuredTemplateConstructorDataData = Parameters<
+  foundry.data.MeasuredTemplateData['_initializeSource']
+>[0];