+++ src/module/config.ts
@@ -1,3 +1,7 @@
+import { TemplateConfig } from '../interfaces/TemplateConfig';
+import SwadeMeasuredTemplate from './documents/SwadeMeasuredTemplate';
+import { TemplatePreset } from './enums/TemplatePresetEnum';
+
 export const SWADE = {
   ASCII: `
   ███████╗██╗    ██╗ █████╗ ██████╗ ███████╗
@@ -262,6 +266,63 @@ export const SWADE = {
     compendium: 'systems/swade/assets/ui/wildcard-dark.svg',
   },
 
+  measuredTemplatePresets: [
+    {
+      data: { t: CONST.MEASURED_TEMPLATE_TYPES.CONE, distance: 9 },
+      button: {
+        name: TemplatePreset.CONE,
+        title: 'SWADE.Cone',
+        icon: 'text-icon cone',
+        visible: true,
+        button: true,
+        onClick: () => {
+          SwadeMeasuredTemplate.fromPreset(TemplatePreset.CONE);
+        },
+      },
+    },
+    {
+      data: { t: CONST.MEASURED_TEMPLATE_TYPES.CIRCLE, distance: 1 },
+      button: {
+        name: TemplatePreset.SBT,
+        title: 'SWADE.SBT',
+        icon: 'text-icon sbt',
+        visible: true,
+        button: true,
+        onClick: () => {
+          SwadeMeasuredTemplate.fromPreset(TemplatePreset.SBT);
+        },
+      },
+    },
+    {
+      data: { t: CONST.MEASURED_TEMPLATE_TYPES.CIRCLE, distance: 2 },
+      button: {
+        name: TemplatePreset.MBT,
+        title: 'SWADE.MBT',
+        icon: 'text-icon mbt',
+        visible: true,
+        button: true,
+        onClick: () => {
+          SwadeMeasuredTemplate.fromPreset(TemplatePreset.MBT);
+        },
+      },
+    },
+    {
+      data: { t: CONST.MEASURED_TEMPLATE_TYPES.CIRCLE, distance: 3 },
+      button: {
+        name: TemplatePreset.LBT,
+        title: 'SWADE.LBT',
+        icon: 'text-icon lbt',
+        visible: true,
+        button: true,
+        onClick: () => {
+          SwadeMeasuredTemplate.fromPreset(TemplatePreset.LBT);
+        },
+      },
+    },
+  ],
+
+  activeMeasuredTemplatePreview: null,
+
   dsnColorSets: {},
 
   dsnTextureList: {},
@@ -270,3 +331,102 @@ export const SWADE = {
 
   allowedActorFlags: [],
 };
+
+export interface SwadeConfig {
+  ASCII: string;
+  attributes: {
+    agility: {
+      long: string;
+      short: string;
+    };
+    smarts: {
+      long: string;
+      short: string;
+    };
+    spirit: {
+      long: string;
+      short: string;
+    };
+    strength: {
+      long: string;
+      short: string;
+    };
+    vigor: {
+      long: string;
+      short: string;
+    };
+  };
+
+  init: {
+    defaultCardCompendium: string;
+    cardTable: string;
+  };
+
+  packChoices: Record<string, string>;
+
+  imagedrop: {
+    height: number;
+  };
+
+  bennies: {
+    templates: {
+      refresh: string;
+      refreshAll: string;
+      add: string;
+      spend: string;
+      gmadd: string;
+      joker: string;
+    };
+  };
+
+  vehicles: {
+    maxHandlingPenalty: number;
+    opSkills: Array<string>;
+  };
+
+  settingConfig: {
+    id: string;
+    title: string;
+    settings: Array<string>;
+  };
+
+  diceConfig: {
+    id: string;
+    title: string;
+    flags: Record<string, any>;
+  };
+
+  actionCardEditor: {
+    id: string;
+  };
+
+  statusEffects: Array<StatusEffect>;
+
+  wildCardIcons: {
+    regular: string;
+    compendium: string;
+  };
+
+  measuredTemplatePresets: Array<TemplateConfig>;
+
+  activeMeasuredTemplatePreview: SwadeMeasuredTemplate | null;
+
+  dsnColorSets: any;
+  dsnTextureList: any;
+
+  raceLocalizationKey: string;
+  allowedActorFlags: Array<string>;
+}
+
+interface StatusEffect {
+  icon: string;
+  id: string;
+  label: string;
+  changes?: Array<StatusEffectChanges>;
+}
+
+interface StatusEffectChanges {
+  key: string;
+  value: string | number;
+  mode: number;
+}