+++ src/module/documents/SwadeUser.ts
@@ -48,8 +48,13 @@ export default class SwadeUser extends User {
       await this.setFlag('swade', 'bennies', this.bennies - 1);
       const dsnShowBennyAnimation = await shouldShowBennyAnimation();
       if (!!game.dice3d && dsnShowBennyAnimation) {
-        const benny = await new Roll('1dB').evaluate();
-        game.dice3d.showForRoll(benny, game.user!, true, null, false);
+        game.dice3d.showForRoll(
+          await new Roll('1dB').evaluate(),
+          game.user!,
+          true,
+          null,
+          false,
+        );
       }
     } else if (this.character) {
       await this.character.spendBenny();